+++ NEWS.md
@@ -1,14 +1,24 @@
+# tvReg 0.4.1
+
+* Added a website for the package (with pkgdown).
+
+* Added 'README.Rmd' and 'index.Rmd'.
+
+* Updated documentation `tvOLS matrix`, `forecast.tvar`, `forecast.tvlm` and
+  `forecast.tvsure`.
+  
+
 # tvReg 0.4.0
 
-* Remove class `tvlm` from objects of class `tvar`.
+* Removed class `tvlm` from objects of class `tvar`.
 
-* Remove deprecated `CI.tvlm()`, `CI.tvar()`, `CI.tvsure()` and `CI.tvirf()` methods
+* Removed deprecated `CI.tvlm()`, `CI.tvar()`, `CI.tvsure()` and `CI.tvirf()` methods.
 
-* `predict.tvlm()`, `predict.tvar()`, `predict.tvsure()` and `predict.tvvar()` methods added
+* Added `predict.tvlm()`, `predict.tvar()`, `predict.tvsure()` and `predict.tvvar()` methods.
 
-* `forecast.tvlm()`, `forecast.tvar()`, `forecast.tvsure()` and `forecast.tvvar()` methods added
+* Added `forecast.tvlm()`, `forecast.tvar()`, `forecast.tvsure()` and `forecast.tvvar()` methods.
 
-* `resid.tvlm()`, `resid.tvsure()`, `resid.tvvar()` and `resid.tvirf()` added for compatility with R standards
+* Added `resid.tvlm()`, `resid.tvsure()`, `resid.tvvar()` and `resid.tvirf()` for compatility with R standards.